# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cargue-masivo/auditorias/plantilla": {
      "get": {
        "operationId": "CargueMasivoController_descargarPlantillaAuditorias",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Plantilla descargada exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "base64": {
                      "type": "string",
                      "description": "Archivo de plantilla en formato Base64."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno."
          }
        },
        "summary": "Descargar plantilla de auditorías",
        "tags": [
          "Cargue Masivo"
        ]
      }
    },
    "/cargue-masivo/auditorias/plan/{planId}": {
      "get": {
        "operationId": "CargueMasivoController_descargarAuditoriasExcel",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "description": "ID del plan de auditoría.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo de auditorías descargado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "base64": {
                      "type": "string",
                      "description": "Archivo de auditorías en formato Base64."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno."
          }
        },
        "summary": "Descargar auditorías en formato Excel",
        "tags": [
          "Cargue Masivo"
        ]
      }
    },
    "/cargue-masivo/auditorias": {
      "post": {
        "operationId": "CargueMasivoController_cargueMasivo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Carga masiva de auditorías con datos en formato Base64.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "base64data": {
                    "type": "string",
                    "description": "Archivo en formato Base64."
                  },
                  "complemento": {
                    "type": "object",
                    "description": "Datos complementarios."
                  }
                },
                "required": [
                  "base64data",
                  "complemento"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Carga masiva exitosa."
          },
          "400": {
            "description": "Datos inválidos."
          },
          "500": {
            "description": "Error interno."
          }
        },
        "summary": "Carga masiva de auditorías",
        "tags": [
          "Cargue Masivo"
        ]
      }
    },
    "/cargue-masivo/actividades": {
      "post": {
        "operationId": "CargueMasivoController_cargueMasivoActividades",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Carga masiva de actividades con datos en formato Base64.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "base64data": {
                    "type": "string",
                    "description": "Archivo en formato Base64."
                  },
                  "complemento": {
                    "type": "object",
                    "description": "Datos complementarios."
                  }
                },
                "required": [
                  "base64data",
                  "complemento"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Carga masiva exitosa."
          },
          "400": {
            "description": "Datos inválidos."
          },
          "500": {
            "description": "Error interno."
          }
        },
        "summary": "Carga masiva de actividades",
        "tags": [
          "Cargue Masivo"
        ]
      }
    },
...
    "/plan-estado": {
      "get": {
        "operationId": "PlanEstadoController_getAll",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filtro opcional.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista obtenida con éxito."
          },
          "404": {
            "description": "No se encontraron estados."
          }
        },
        "summary": "Obtiene todos los estados de planes.",
        "tags": [
          "Plan Estado"
        ]
      }
    },
    "/plan-estado/{id}": {
      "get": {
        "operationId": "PlanEstadoController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del estado del plan.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado obtenido con éxito."
          },
          "404": {
            "description": "No se encontró el estado."
          }
        },
        "summary": "Obtiene un estado de plan por ID.",
        "tags": [
          "Plan Estado"
        ]
      }
    },
    "/plantilla/{id}": {
      "get": {
        "operationId": "PlantillaController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la plantilla",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conEspeciales",
            "required": false,
            "in": "query",
            "description": "Incluir caracteres especiales en la plantilla",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "auditoria-padre",
            "required": false,
            "in": "query",
            "description": "Usar la auditoría padre en la plantilla en lugar de la colección de auditorías original",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plantilla encontrada."
          }
        },
        "summary": "Obtener plantilla por ID",
        "tags": [
          "Plantilla"
        ]
      }
    },
    "/plantilla/{tipo}/{idAuditoria}": {
      "get": {
        "operationId": "PlantillaController_getPlantilla",
        "parameters": [
          {
            "name": "tipo",
            "required": true,
            "in": "path",
            "description": "Tipo de la plantilla (ej: plan-trabajo, solicitud-informacion, carta-presentacion, programa-trabajo, informe-seguimiento, informe-auditoria)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idAuditoria",
            "required": true,
            "in": "path",
            "description": "ID de la auditoría",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plantilla generada exitosamente."
          },
          "404": {
            "description": "Tipo de plantilla no encontrado."
          }
        },
        "summary": "Obtener plantilla dinámica según el tipo de auditoría",
        "tags": [
          "Plantilla"
        ]
      }
    },
    "/plantilla/docx/{tipo}/{idAuditoria}": {
      "get": {
        "operationId": "PlantillaController_getPlantillaDOCX",
        "parameters": [
          {
            "name": "tipo",
            "required": true,
            "in": "path",
            "description": "Tipo de la plantilla (ej: plan-trabajo, solicitud-informacion, carta-presentacion, programa-trabajo, informe-seguimiento, informe-auditoria)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idAuditoria",
            "required": true,
            "in": "path",
            "description": "ID de la auditoría",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plantilla generada exitosamente."
          },
          "404": {
            "description": "Tipo de plantilla no encontrado."
          }
        },
        "summary": "Obtener plantilla dinámica según el tipo de auditoría en formato DOCX",
        "tags": [
          "Plantilla"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan Anual Auditoria MID",
    "description": "API MID para la gestion de planes de auditorias, auditorias y actividades",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}